refactor: change root structure accordingly react navigation scheme
