--- app/app.tsx
@@ -1,17 +1,0 @@
-import { Text, View } from 'react-native';
-import LoginScreen from './screens/login/login';
-
-export default function App() {
-  return (
-    <View
-      style={{
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center',
-      }}
-    >
-      <LoginScreen/>
-      <Text>This is App text placeholder</Text>
-    </View>
-  );
-}
