--- app/index.tsx
@@ -1,24 +1,0 @@
-import { View } from 'react-native';
-import {
-  QueryClient,
-  QueryClientProvider,
-} from '@tanstack/react-query';
-import App from './app';
-
-export default function Index() {
-  const queryClient = new QueryClient();
-
-  return (
-    <View
-      style={{
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center',
-      }}
-    >
-      <QueryClientProvider client={queryClient}>
-        <App />
-      </QueryClientProvider>
-    </View>
-  );
-}
