--- expo-env.d.ts
@@ -1,3 +1,0 @@
-/// <reference types="expo/types" />
-
-// NOTE: This file should not be edited and should be in your git ignore
