+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-native-music-app",
-  "main": "expo-router/entry",
+  "main": "expo/AppEntry.js",
   "version": "1.0.0",
   "scripts": {
     "start": "expo start",
@@ -20,6 +20,7 @@
     "@expo/vector-icons": "^14.0.2",
     "@react-native-async-storage/async-storage": "^1.24.0",
     "@react-navigation/native": "^6.1.18",
+    "@react-navigation/native-stack": "^6.11.0",
     "@reduxjs/toolkit": "^2.2.7",
     "@spotify/web-api-ts-sdk": "^1.2.0",
     "@tanstack/react-query": "^5.51.21",
